feat: add, edit & delete functions for dues & events services

+++ app/src/main/java/com/collogestion/data/Due.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 class Due (
     @SerializedName("id") var id : Int,
-    @SerializedName("amount") var amount : Float,
+    @SerializedName("amount") var amount : Double,
     @SerializedName("creditor_id") var creditorId : Int,
     @SerializedName("debtor_id") var debtorId : Int,
     @SerializedName("house_share_id") var houseShareId : Int
